Made some changes to make v4.2.x compile for vxWorks

+++ CoolProp/FluidClass.cpp
@@ -732,7 +732,7 @@ double Fluid::density_Tp(double T, double p)
 
 double Fluid::density_Tp(double T, double p, double rho_guess)
 {
-    long double tau,dpdrho__constT,dpddelta__constT, error=999,R,p_EOS,rho,change=999;
+    double tau,dpdrho__constT,dpddelta__constT, error=999,R,p_EOS,rho,change=999;
 
     R = params.R_u/params.molemass*1000; // SI units are used internally
 	tau = reduce.T/T;
@@ -745,13 +745,13 @@ double Fluid::density_Tp(double T, double p, double rho_guess)
 	// In subsequent steps, use secant method because each evaluation of newton step requires two evaluations of derivatives with respect to delta
 	rho=rho_guess;
     int iter=1;
-	long double delta = rho/reduce.rho;
+	double delta = rho/reduce.rho;
 	
 	while (std::abs(error) > 1e-9 && std::abs(change) > 1e-13) 
     {
 		// Needed for both p and p derivative
 		// Run once to cut down on calculations
-		long double dphirdDelta = dphir_dDelta(tau, delta);
+		double dphirdDelta = dphir_dDelta(tau, delta);
 
 		// Pressure from equation of state
 		p_EOS = reduce.rho*delta*R*T*(1+delta*dphirdDelta);
@@ -1149,8 +1149,8 @@ void Fluid::rhosatPure_Akasaka(double T, double &rhoLout, double &rhoVout, doubl
 
 	Ancillary equations are used to get a sensible starting point
 	*/
-	long double rhoL,rhoV,dphirL,dphirV,ddphirL,ddphirV,phirL,phirV,JL,JV,KL,KV,dJL,dJV,dKL,dKV;
-	long double DELTA, deltaL=0, deltaV=0, tau=0, error, PL, PV, stepL, stepV;
+	double rhoL,rhoV,dphirL,dphirV,ddphirL,ddphirV,phirL,phirV,JL,JV,KL,KV,dJL,dJV,dKL,dKV;
+	double DELTA, deltaL=0, deltaV=0, tau=0, error, PL, PV, stepL, stepV;
 	int iter=0;
 	// Use the density ancillary function as the starting point for the solver
     try
